Split dependency injection extension classes for file system, media and CDN
