+++ src/Umbraco.StorageProviders.AzureBlob/DependencyInjection/AzureBlobMediaFileSystemExtensions.cs
@@ -1,0 +1,130 @@
+﻿using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Options;
+using SixLabors.ImageSharp.Web.Caching;
+using SixLabors.ImageSharp.Web.Providers;
+using Umbraco.Cms.Core.Configuration.Models;
+using Umbraco.Cms.Infrastructure.DependencyInjection;
+using Umbraco.Cms.Web.Common.ApplicationBuilder;
+using Umbraco.Extensions;
+using Umbraco.StorageProviders.AzureBlob;
+using Umbraco.StorageProviders.AzureBlob.Imaging;
+using Umbraco.StorageProviders.AzureBlob.IO;
+
+// ReSharper disable once CheckNamespace
+// uses same namespace as Umbraco Core for easier discoverability
+namespace Umbraco.Cms.Core.DependencyInjection
+{
+    /// <summary>
+    /// Extension methods to help registering Azure Blob Storage file systems for Umbraco media.
+    /// </summary>
+    public static class AzureBlobMediaFileSystemExtensions
+    {
+        /// <summary>
+        /// Registers a <see cref="IAzureBlobFileSystem"/> and it's dependencies configured for media.
+        /// </summary>
+        /// <param name="builder">The <see cref="IUmbracoBuilder"/>.</param>
+        /// <returns>The <see cref="IUmbracoBuilder"/>.</returns>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="builder"/> is null.</exception>
+        public static IUmbracoBuilder AddAzureBlobMediaFileSystem(this IUmbracoBuilder builder)
+        {
+            if (builder == null) throw new ArgumentNullException(nameof(builder));
+
+            builder.AddAzureBlobFileSystem(AzureBlobFileSystemOptions.MediaFileSystemName, "~/media",
+                (options, provider) =>
+                {
+                    var globalSettingsOptions = provider.GetRequiredService<IOptions<GlobalSettings>>();
+                    options.VirtualPath = globalSettingsOptions.Value.UmbracoMediaPath;
+                });
+
+            builder.Services.TryAddSingleton<AzureBlobMediaMiddleware>();
+
+            builder.Services.AddUnique<IImageProvider, AzureBlobImageProvider>();
+            builder.Services.AddUnique<IImageCache>(provider => new AzureBlobImageCache(provider
+                    .GetRequiredService<IOptionsFactory<AzureBlobFileSystemOptions>>()
+                    .Create(AzureBlobFileSystemOptions.MediaFileSystemName)
+            ));
+
+            builder.SetMediaFileSystem(provider => provider.GetRequiredService<IAzureBlobFileSystemProvider>()
+                .GetFileSystem(AzureBlobFileSystemOptions.MediaFileSystemName));
+
+            return builder;
+        }
+
+        /// <summary>
+        /// Registers a <see cref="IAzureBlobFileSystem"/> and it's dependencies configured for media.
+        /// </summary>
+        /// <param name="builder">The <see cref="IUmbracoBuilder"/>.</param>
+        /// <param name="configure">An action used to configure the <see cref="AzureBlobFileSystemOptions"/>.</param>
+        /// <returns>The <see cref="IUmbracoBuilder"/>.</returns>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="builder"/> or <paramref name="configure"/> is null.</exception>
+        public static IUmbracoBuilder AddAzureBlobMediaFileSystem(this IUmbracoBuilder builder, Action<AzureBlobFileSystemOptions> configure)
+        {
+            if (builder == null) throw new ArgumentNullException(nameof(builder));
+            if (configure == null) throw new ArgumentNullException(nameof(configure));
+
+            AddAzureBlobMediaFileSystem(builder);
+
+            builder.Services
+                .AddOptions<AzureBlobFileSystemOptions>(AzureBlobFileSystemOptions.MediaFileSystemName)
+                .Configure(configure);
+
+            return builder;
+        }
+
+        /// <summary>
+        /// Registers a <see cref="IAzureBlobFileSystem"/> and it's dependencies configured for media.
+        /// </summary>
+        /// <param name="builder">The <see cref="IUmbracoBuilder"/>.</param>
+        /// <param name="configure">An action used to configure the <see cref="AzureBlobFileSystemOptions"/>.</param>
+        /// <returns>The <see cref="IUmbracoBuilder"/>.</returns>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="builder"/> or <paramref name="configure"/> is null.</exception>
+        public static IUmbracoBuilder AddAzureBlobMediaFileSystem(this IUmbracoBuilder builder, Action<AzureBlobFileSystemOptions, IServiceProvider> configure)
+        {
+            if (builder == null) throw new ArgumentNullException(nameof(builder));
+            if (configure == null) throw new ArgumentNullException(nameof(configure));
+
+            AddAzureBlobMediaFileSystem(builder);
+
+            builder.Services
+                .AddOptions<AzureBlobFileSystemOptions>(AzureBlobFileSystemOptions.MediaFileSystemName)
+                .Configure(configure);
+
+            return builder;
+        }
+
+        /// <summary>
+        /// Adds the <see cref="AzureBlobMediaMiddleware" />.
+        /// </summary>
+        /// <param name="builder">The <see cref="IUmbracoApplicationBuilderContext" />.</param>
+        /// <returns>
+        /// The <see cref="IUmbracoApplicationBuilderContext" />.
+        /// </returns>
+        /// <exception cref="System.ArgumentNullException">builder</exception>
+        public static IUmbracoApplicationBuilderContext UseAzureBlobMediaFileSystem(this IUmbracoApplicationBuilderContext builder)
+        {
+            if (builder == null) throw new ArgumentNullException(nameof(builder));
+
+            UseAzureBlobMediaFileSystem(builder.AppBuilder);
+
+            return builder;
+        }
+
+        /// <summary>
+        /// Adds the <see cref="AzureBlobMediaMiddleware"/>.
+        /// </summary>
+        /// <param name="app">The <see cref="IApplicationBuilder"/>.</param>
+        /// <returns>The <see cref="IApplicationBuilder"/>.</returns>
+        /// <exception cref="ArgumentNullException">Thrown if <paramref name="app"/> or is null.</exception>
+        public static IApplicationBuilder UseAzureBlobMediaFileSystem(this IApplicationBuilder app)
+        {
+            if (app == null) throw new ArgumentNullException(nameof(app));
+
+            app.UseMiddleware<AzureBlobMediaMiddleware>();
+
+            return app;
+        }
+    }
+}
